+++ src/firebase/auth/index.ts
@@ -1,4 +1,4 @@
-import { getAuth } from 'firebase/auth';
+import { getAuth } from '@firebase/auth';
 import { app } from '../app';
 
 export const auth = getAuth(app);
