+++ src/firebase/auth/useSignIn.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback, useMemo } from 'react';
-import { signInWithEmailAndPassword } from 'firebase/auth';
+import { signInWithEmailAndPassword } from '@firebase/auth';
 import { auth } from './index';
 
 /* -------------------------------------------------------------------------- */
