+++ src/firebase/storage/useDownloadURL.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect, useMemo } from 'react';
-import { getDownloadURL, listAll, ref } from 'firebase/storage';
+import { getDownloadURL, listAll, ref } from '@firebase/storage';
 import { storage } from './index';
 
 /* -------------------------------------------------------------------------- */
